requirements.txt
torch
torchvision
numpy
pillow
requests

matplotlib
pandas
opencv-python